+++ Packages/com.adapty.unity-sdk/Runtime/Models/AdaptyInstallationStatusType.cs
@@ -1,0 +1,21 @@
+//
+//  AdaptyInstallationStatusType.cs
+//  AdaptySDK
+//
+
+using UnityEngine.Scripting;
+using System.Runtime.Serialization;
+
+namespace AdaptySDK
+{
+    [Preserve]
+    public enum AdaptyInstallationStatusType
+    {
+        [EnumMember(Value = "not_available")]
+        NotAvailable = 0,
+        [EnumMember(Value = "not_determined")]
+        NotDetermined = 1,
+        [EnumMember(Value = "determined")]
+        Determined = 2,
+    }
+}